Betting and simple betting list in!

+++ src/main/java/pl/coderslab/sports_betting/Entity/User.java
@@ -4,6 +4,7 @@ import lombok.Data;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotEmpty;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
@@ -34,7 +35,7 @@ public @Data class User {
 
     private boolean enabled;
 
-    private BigInteger money;
+    private BigDecimal money;
 
     @ManyToMany(cascade = CascadeType.ALL, fetch = FetchType.EAGER)
     @JoinTable(name = "users_roles", joinColumns = @JoinColumn(name = "user_id"),
